refactor: 클래스 이름 및 구조 변경
