+++ src/main/java/lotto/Application.java
@@ -1,7 +1,7 @@
 package lotto;
 
 import lotto.model.CheckResults;
-import lotto.view.Input;
+import lotto.view.InputView;
 import lotto.view.InputValidator;
 import lotto.model.TicketGenerator;
 
@@ -10,19 +10,19 @@ import java.util.List;
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
-        Input input = new Input();
+        InputView inputView = new InputView();
 
 
-        long money = InputValidator.validatedMoney(input.enterMoney());
+        long money = InputValidator.validatedMoney(inputView.enterMoney());
         System.out.println("구매금액 : " + money); // 테스트 출력. long 자료형
 
         TicketGenerator ticketGenerator = new TicketGenerator(money);
         System.out.println(ticketGenerator.getTickets()); // 테스트 출력
 
-        List<Integer> lottoNumbers = InputValidator.validatedLottoNumbers(input.enterLottoNumbers());
+        List<Integer> lottoNumbers = InputValidator.validatedLottoNumbers(inputView.enterLottoNumbers());
         System.out.println("당첨번호 : " + lottoNumbers); // 테스트 출력. List<Integer>
 
-        int bonusNumber = InputValidator.validatedBonusNumber(input.enterBonusNumber());
+        int bonusNumber = InputValidator.validatedBonusNumber(inputView.enterBonusNumber());
         System.out.println("보너스번호 : " + bonusNumber); // 테스트 출력
 
         CheckResults checkResults = new CheckResults(lottoNumbers, bonusNumber, ticketGenerator.getTickets());
