+++ src/main/java/lotto/model/Bonus.java
@@ -1,6 +1,5 @@
-package lotto.view;
+package lotto.model;
 
-import java.util.Collections;
 import java.util.List;
 
 public class Bonus {
