+++ src/main/java/lotto/model/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.view;
+package lotto.model;
 
 import java.util.Collections;
 import java.util.HashSet;
