+++ src/main/java/lotto/view/InputView.java
@@ -2,7 +2,7 @@ package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
 
-public class Input {
+public class InputView {
 
     static final String PROMPT_MONEY = "구입금액을 입력해 주세요.";
     static final String PROMPT_LOTTO_NUMBERS = "당첨 번호를 입력해 주세요.";
